Refactoring huggingface.py and modelscope.py and extract repo_model_base.py
